requirements.txt
openai
streamlit
dotenv
gtts
google-api-python-client
google-auth-httplib2
google-auth-oauthlib